# backend/src/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Akounamatata",
    "version": "1.0.0",
    "description": "API RESTful pour le système de restaurant digital Akounamatata",
    "contact": {
      "name": "Équipe Akounamatata",
      "email": "contact@akounamatata.ci"
    },
    "license": {
      "name": "ISC",
      "url": "https://opensource.org/licenses/ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Serveur de développement"
    },
    {
      "url": "https://api.akounamatata.ci",
      "description": "Serveur de production"
    }
  ],
  "paths": {
...
    "/api/table/{id}/occupy": {
      "post": {
        "tags": ["Tables"],
        "summary": "Occuper une table",
        "description": "Marque une table comme occupée",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID de la table",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Table occupée",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "table": {
                          "$ref": "#/components/schemas/Table"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Table occupée"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Table déjà occupée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Table non trouvée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Erreur interne du serveur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/table/{id}/free": {
      "post": {
        "tags": ["Tables"],
        "summary": "Libérer une table",
        "description": "Marque une table comme libre",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID de la table",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Table libérée",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "table": {
                          "$ref": "#/components/schemas/Table"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Table libérée"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Table déjà libre",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Table non trouvée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Erreur interne du serveur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Table": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID unique de la table"
          },
          "numero": {
            "type": "string",
            "description": "Numéro de la table"
          },
          "capacite": {
            "type": "number",
            "description": "Capacité de la table"
          },
          "statutOccupee": {
            "type": "boolean",
            "description": "Statut d'occupation de la table"
          },
          "clientActuel": {
            "type": "string",
            "description": "ID du client actuel (si occupée)"
          },
          "qrCode": {
            "type": "string",
            "description": "Code QR de la table"
          },
          "restaurantId": {
            "type": "string",
            "description": "ID du restaurant"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "data": {
            "type": "object",
            "properties": {
              "errors": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "field": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "message": {
            "type": "string",
            "description": "Message d'erreur"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}